+++ src/main/java/com/asapp/backend/challenge/application/model/responses/CreateMessageResponse.java
@@ -4,15 +4,15 @@ import java.util.Date;
 
 public class CreateMessageResponse {
 
-    private final int id;
+    private final long id;
     private final Date timestamp;
 
-    public CreateMessageResponse(int id, Date timestamp) {
+    public CreateMessageResponse(long id, Date timestamp) {
         this.id = id;
         this.timestamp = timestamp;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
